Move ContentPane to adaptive package
ContentPane is an adaptive layout concern, not a scaffold component.
