+++ lib/design/fe/src/commonMain/kotlin/cz/adamec/timotej/snag/lib/design/fe/scenes/MapListDetailSceneStrategy.kt
@@ -41,10 +41,10 @@ import androidx.navigation3.runtime.NavEntry
 import androidx.navigation3.scene.Scene
 import androidx.navigation3.scene.SceneStrategy
 import androidx.navigation3.scene.SceneStrategyScope
+import cz.adamec.timotej.snag.lib.design.fe.adaptive.ContentPane
+import cz.adamec.timotej.snag.lib.design.fe.adaptive.ContentPaneSpacing
 import cz.adamec.timotej.snag.lib.design.fe.adaptive.isScreenExtraWide
 import cz.adamec.timotej.snag.lib.design.fe.adaptive.isScreenWide
-import cz.adamec.timotej.snag.lib.design.fe.scaffold.ContentPane
-import cz.adamec.timotej.snag.lib.design.fe.scaffold.ContentPaneSpacing
 
 class MapListDetailSceneStrategy<T : Any> : SceneStrategy<T> {
     @Suppress("ReturnCount")
